Change API for setting the request environment (#11)
* Improve tests for Configuration#url

Clarify some the test descriptions, and add a test case for a missed
edge case, when the test mode setting is not explicitly set.

* Change API for setting request environment

* Change default request environment to sandbox

This is done for a generally safer library that requires explicit
configuration to access the production environment.

* Raise error when environment is set incorrectly

* Support environment argument in Configuration#url

This is an override argument, which will allow overriding the static
environment configuration.

* Support per-request `:environment` configuration